EDWIN_Discord/discord_bot/app/requirements.txt
discord.py
python-dotenv
openai
requests
google-cloud-speech
google-cloud-translate
azure-cognitiveservices-speech
pydub
numpy